Change how stir(x) behaves
Arnold had `x = x/256;` in his stir(x) function. However, when using `evt.when` for each key press
(down and up), which is a millisecond timestamp, this meant really only mixing the RC4 state 4 (old)
or 5 (recent) times, which isn't variable at all. Instead, I take the millisecond number only
(0-999) and mix the Spritz state that many times for every key press event. This is actually highly
variable and provides greater confidence in the security of the generated passwords.

Also due to Spectre, `window.performance.now()` is no longer reliable as a microsecond-precise
timestamp. Firefox reduces its precision to 1 millisecond as an anti-fingerprinting technique, and
other browsers could reduce it further to 100 milliseconds. So instead, it's back to relying on
`Date.now()` which is reliable across all browsers, but only 1 millisecond precise.