Change output character design
The problem with outputting 1 character for every 3 characters typed is getting
rewarded differently based on the chosen template. For example, the
entropy-per-character of base-6 "Senary" is approximately 2.58 bits, but the
entropy-per-character of base-94 "ASCII" is approximatly 6.55 bits. So for a 128
minimum bit security margin, you will type only 60 characters for an ASCII
password but 150 characters for a Senary one.

The new approach awards the number of bits per character typed, currently set to
2. As such, a user will only type on average about 70 characters for a
20-character ASCII password, and 75 characters for a 50-character Senary one.
This is far more efficient.